# PERF: assert_frame_equal and assert_series_equal for frames/series with a MultiIndex